Add completions for message and service request  content
Once a topic name and message type (or service name and type) have been entered in the command line, pressing TAB will insert a string with all fields of the message and their default values, making it easy to publish even complex messages.
Note that below fish version 4.x, this will insert an ugly string with lots of escaping backlashes. As it doesn't really break anything, I didn't change the requirement to be higher than 4, only removed an error when using a larger major version with the minor being smaller than 6.